+++ packages/testing/src/chip/chip.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { globSync } from "@matter/tools";
+import { globSync } from "@nacho-iot/js-tools";
 import { Subject } from "../device/subject.js";
 import { Test } from "../device/test.js";
 import type { Container } from "../docker/container.js";