+++ support/codegen/src/forwards/matterjs-main.ts
@@ -5,7 +5,7 @@
  */
 
 import { TsFile } from "#util/TsFile.js";
-import { Graph } from "@matter/tools";
+import { Graph } from "@nacho-iot/js-tools";
 import { generateForwards } from "./generate-forwards.js";
 
 /**